commit something

+++ run.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 pid=`ps aux | grep java | grep play | awk '{print $2}'`
-kill -9 $pid
-echo $pid
+if [ -n $pid ]; then
+ kill -9 $pid
+fi
+echo "pid" $pid
 play run --%uttp | /etc/colorlog
